+++ src/docfx/docfx.csproj
@@ -40,19 +40,9 @@
   <ItemGroup>
     <ProjectReference Include="../Microsoft.Docs.Template/Microsoft.Docs.Template.csproj" />
   </ItemGroup>
-
   <ItemGroup>
-    <None Update="data\ContextObject.json">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </None>
-    <None Update="data\LandingData.json">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </None>
-    <None Update="data\TestData.json">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </None>
-    <None Update="data\TestPage.json">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    <None Update="data/*.json">
+        <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
   </ItemGroup>
